done basic all api of expanse tracker

+++ src/main/java/internship/intern/dto/ExpanseDTO.java
@@ -17,7 +17,9 @@ public class ExpanseDTO {
 	
 	private LocalDate date;
 	
-	private Integer amount;
+	private Double amount;
+
+
